rename fixed file
